requirements.txt
certifi==2018.11.29
chardet==3.0.4
Django==2.1
django-shortuuidfield==0.1.3
furl==1.0.1
idna==2.6
orderedmultidict==1.0
Pillow==5.3.0
PyMySQL==0.8.0
python-dateutil==2.7.5
pytz==2018.7
redis==2.10.6
requests==2.18.4
shortuuid==0.5.0
six==1.12.0
TimeConvert==1.4.3
tzlocal==1.5.1
urllib3==1.22
xlrd==1.2.0
